rename ContainerCluster file name
